query refactoring
